layout/mathml/tests/test_bug1845461.html

<!DOCTYPE HTML>
<!--
    https://bugzilla.mozilla.org/show_bug.cgi?id=1845461
  -->
<html>
  <head>
    <title>Test for Bug 1845461</title>
    <script src="/tests/SimpleTest/SimpleTest.js"></script>
    <script src="/tests/SimpleTest/EventUtils.js"></script>
    <link rel="stylesheet" type="text/css" href="/tests/SimpleTest/test.css" />
    <script type="application/javascript">
      let stringBundleService =
          SpecialPowers.Cc["@mozilla.org/intl/stringbundle;1"]
          .getService(SpecialPowers.Ci.nsIStringBundleService);
      const allow_legacy_mathvariant = !SpecialPowers.getBoolPref('mathml.legacy_mathvariant_attribute.disabled');
      let errorMessages = [];
      SpecialPowers.registerConsoleListener(
          function (msg) {
              if (msg.message == "SENTINEL") {
                  is(errorMessages.length, allow_legacy_mathvariant ? 1 : 0,
                     "number of errors");
                  if (allow_legacy_mathvariant) {
                      is(errorMessages[0], "“mathvariant='fraktur'” on MathML elements is deprecated and will be removed at a future date.",
                        "error message");
                  }
                  SimpleTest.finish();
              } else if (msg.isScriptError) {
                  errorMessages.push(msg.errorMessage);
              }
          });
      SimpleTest.waitForExplicitFinish();
    </script>
  </head>
  <body onload="SpecialPowers.postConsoleSentinel();">
    <a target="_blank" href="https://bugzilla.mozilla.org/show_bug.cgi?id=1845461">Mozilla Bug 1845461</a>
    <math><mi mathvariant="normal">A</mi></math>
    <math><mi mathvariant="fraktur">A</mi></math>
  </body>
</html>
